+++ packages/themes/src/composables/use-theme.ts
@@ -5,8 +5,8 @@ import type { MaybeRef } from 'vue';
 import { computed, unref } from 'vue';
 import type { Theme } from '../schemas/theme.js';
 import { useThemeStore } from '../stores/theme.js';
-import { theme as themeDefaultDark } from '../themes/dark/directus-default.js';
-import { theme as themeDefaultLight } from '../themes/light/directus-default.js';
+import { theme as themeDefaultDark } from '../themes/dark/default.js';
+import { theme as themeDefaultLight } from '../themes/light/default.js';
 
 export const useTheme = (
 	darkMode: MaybeRef<boolean>,
