+++ packages/themes/src/themes/dark/index.ts
@@ -1,1 +1,1 @@
-export { theme as directusDefault } from './directus-default.js';
+export { default as directusDefault } from './default.js';
