+++ packages/themes/src/themes/light/color-match.ts
@@ -1,6 +1,6 @@
-import type { Theme } from '../../schemas/theme.js';
+import { defineTheme } from '../../index.js';
 
-export const theme: Theme = {
+export default defineTheme({
 	name: 'Directus Color Match',
 	appearance: 'light',
 	rules: {
@@ -74,4 +74,4 @@ export const theme: Theme = {
 			background: '#FFFFFF',
 		},
 	},
-};
+});
