+++ packages/themes/src/themes/light/minimal.ts
@@ -1,6 +1,6 @@
-import type { Theme } from '../../schemas/theme.js';
+import { defineTheme } from '../../index.js';
 
-export const theme: Theme = {
+export default defineTheme({
 	name: 'Directus Minimal',
 	appearance: 'light',
 	rules: {
@@ -11,4 +11,4 @@ export const theme: Theme = {
 			},
 		},
 	},
-};
+});
